Part 1 - Mechanical split into multiple providers
