+++ src/contexts/index.ts
@@ -1,2 +1,4 @@
 export * from './ThemeProvider'
 export * from './TreeStateProvider'
+export * from './EditingProvider'
+export * from './CollapseProvider'
