+++ src/hooks/index.ts
@@ -2,3 +2,4 @@ export * from './useCommon'
 export * from './useDragNDrop'
 export * from './useCollapseTransition'
 export * from './useTriggers'
+export * from './DragSourceProvider'
